refactor: move extensions to Supplements folder and organize code
Move UIViewController, UIColor, UIStackView, and UITextField extensions
from Utilities/Extensions to Supplements to improve project structure.
Split UIViewController extensions into separate files by functionality:
isIphone check, alert display, and loader animations.
Add a convenience initializer for UIStackView to simplify setup.
This reorganization enhances code maintainability and clarity.
