+++ QuickRate/QuickRate/Supplements/UIColor.hex.swift
